restrict comment deletion to own comments

+++ app/controllers/comments_controller.rb
@@ -1,5 +1,4 @@
 class CommentsController < ApplicationController
-
   def create
     @comment = current_user.comments.build(comment_params)
 
@@ -12,6 +11,10 @@ class CommentsController < ApplicationController
 
   def destroy
     @comment = Comment.find(params[:id])
+
+    if @comment.user != current_user
+      return redirect_to post_path(@comment.post), alert: "You can only delete your own comments."
+    end
     
     if @comment.destroy
       redirect_to post_path(@comment.post), notice: "Comment was successfully deleted."
